+++ docs/changelog.md
@@ -9,6 +9,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 # Unreleased
 
+- Fix broken `__usage_on_types__`. Replaced by `__default_usage_on_types__`.
 - Gently raise connection errors.
 - Warn on possible typo in config key.